# src/openapi.json
{
  "openapi": "3.1.0",
  "info": {"title": "급식 배틀 Backend API", "version": "1.0.0"},
  "servers": [{"url": "/api/v1"}],
  "paths": {
    "/schools": {
      "get": {
        "operationId": "searchSchools",
        "parameters": [
          {"name": "query", "in": "query", "required": true, "schema": {"type": "string", "minLength": 2}},
          {"name": "limit", "in": "query", "required": false, "schema": {"type": "integer", "default": 20, "minimum": 1, "maximum": 20}}
        ],
        "responses": {
          "200": {"description": "학교 검색 결과", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/SchoolSearchResponse"}}}},
          "422": {"$ref": "#/components/responses/Problem"},
          "502": {"$ref": "#/components/responses/Problem"},
          "504": {"$ref": "#/components/responses/Problem"},
          "500": {"$ref": "#/components/responses/Problem"}
        }
      }
    },
    "/meals": {
      "get": {
        "operationId": "getMeals",
        "parameters": [
          {"name": "officeCode", "in": "query", "required": true, "schema": {"type": "string", "minLength": 1, "pattern": "^\\S+$"}},
          {"name": "schoolCode", "in": "query", "required": true, "schema": {"type": "string", "minLength": 1, "pattern": "^\\S+$"}},
          {"name": "from", "in": "query", "required": true, "schema": {"type": "string", "format": "date"}},
          {"name": "to", "in": "query", "required": true, "schema": {"type": "string", "format": "date"}}
        ],
        "responses": {
          "200": {"description": "급식 조회 결과", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/MealSearchResponse"}}}},
          "422": {"$ref": "#/components/responses/Problem"},
          "502": {"$ref": "#/components/responses/Problem"},
          "504": {"$ref": "#/components/responses/Problem"},
          "500": {"$ref": "#/components/responses/Problem"}
        }
      }
    },
    "/health": {
      "get": {
        "operationId": "getHealth",
        "responses": {"200": {"description": "정상", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HealthResponse"}}}}}
      }
    }
  },
  "components": {
    "responses": {
      "Problem": {"description": "Problem Details", "content": {"application/problem+json": {"schema": {"$ref": "#/components/schemas/ProblemDetail"}}}}
    },
    "schemas": {
      "SchoolSummary": {
        "type": "object", "additionalProperties": false,
        "required": ["officeCode", "schoolCode", "name", "schoolType", "region", "address"],
        "properties": {
          "officeCode": {"type": "string"}, "schoolCode": {"type": "string"}, "name": {"type": "string"},
          "schoolType": {"type": ["string", "null"]}, "region": {"type": ["string", "null"]}, "address": {"type": ["string", "null"]}
        }
      },
      "SchoolSearchResponse": {
        "type": "object", "additionalProperties": false, "required": ["items", "hasMore"],
        "properties": {"items": {"type": "array", "maxItems": 20, "items": {"$ref": "#/components/schemas/SchoolSummary"}}, "hasMore": {"type": "boolean"}}
      },
      "SelectedSchool": {
        "type": "object", "additionalProperties": false, "required": ["officeCode", "schoolCode", "name"],
        "properties": {"officeCode": {"type": "string"}, "schoolCode": {"type": "string"}, "name": {"type": "string"}}
      },
      "MealItem": {
        "type": "object", "additionalProperties": false, "required": ["date", "mealType", "dishes", "calories"],
        "properties": {
          "date": {"type": "string", "format": "date"}, "mealType": {"type": "string", "const": "중식"},
          "dishes": {"type": "array", "minItems": 1, "items": {"type": "string", "minLength": 1}},
          "calories": {"type": ["string", "null"]}
        }
      },
      "MealSearchResponse": {
        "type": "object", "additionalProperties": false, "required": ["school", "from", "to", "items"],
        "properties": {
          "school": {"$ref": "#/components/schemas/SelectedSchool"}, "from": {"type": "string", "format": "date"},
          "to": {"type": "string", "format": "date"}, "items": {"type": "array", "items": {"$ref": "#/components/schemas/MealItem"}}
        }
      },
      "HealthResponse": {
        "type": "object", "additionalProperties": false, "required": ["status"],
        "properties": {"status": {"type": "string", "const": "ok"}}
      },
      "ProblemDetail": {
        "type": "object", "additionalProperties": false,
        "required": ["type", "title", "status", "detail", "code", "instance"],
        "properties": {
          "type": {"type": "string", "format": "uri"}, "title": {"type": "string"}, "status": {"type": "integer"},
          "detail": {"type": "string"}, "code": {"type": "string"},
          "instance": {"type": ["string", "null"], "format": "uri"}
        }
      }
    }
  }
}